Moved components into new folder for ease of organizing.
