+++ client/app.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import SimilarListView from './SimilarListView';
+import SimilarListView from './components/SimilarListView';
 
 
 ReactDOM.render(<SimilarListView />, document.getElementById('app'));
