+++ client/components/ThumbnailView.jsx
